+++ next-cloudinary/src/components/CldUploadWidget/index.ts
@@ -1,8 +1,2 @@
 export { default } from './CldUploadWidget';
-export type {
-  CldUploadWidgetInfo,
-  CldUploadWidgetProps,
-  CldUploadWidgetPropsChildren,
-  CldUploadWidgetPropsOptions,
-  CldUploadWidgetResults,
-} from './CldUploadWidget.types';
+export type { CldUploadWidgetProps, CldUploadWidgetPropsChildren } from './CldUploadWidget.types';