+++ next-cloudinary/src/helpers/getCldVideoUrl.ts
@@ -1,7 +1,7 @@
 import { constructCloudinaryUrl } from '@cloudinary-util/url-loader';
 import type { VideoOptions, ConfigOptions, AnalyticsOptions } from '@cloudinary-util/url-loader';
 
-import { NEXT_CLOUDINARY_ANALYTICS_ID, NEXT_CLOUDINARY_VERSION, NEXT_VERSION } from '../constants/analytics';
+import { NEXT_CLOUDINARY_ANALYTICS_PRODUCT_ID, NEXT_CLOUDINARY_ANALYTICS_ID, NEXT_CLOUDINARY_VERSION, NEXT_VERSION } from '../constants/analytics';
 import {checkForCloudName} from "../lib/cloudinary";
 
 /**
@@ -13,6 +13,7 @@ export type GetCldVideoUrlConfig = ConfigOptions;
 export type GetCldVideoUrlAnalytics = AnalyticsOptions;
 
 export function getCldVideoUrl(options: GetCldVideoUrlOptions, config?: GetCldVideoUrlConfig, analytics?: GetCldVideoUrlAnalytics) {
+  // @ts-expect-error Property 'cloud' does not exist on type 'CloudinaryConfigurationOptions'.
   const cloudName = config?.cloud?.cloudName ?? process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME;
   checkForCloudName(cloudName);
   return constructCloudinaryUrl({
@@ -26,6 +27,7 @@ export function getCldVideoUrl(options: GetCldVideoUrlOptions, config?: GetCldVi
       },
     }, config),
     analytics: Object.assign({
+      product: NEXT_CLOUDINARY_ANALYTICS_PRODUCT_ID,
       sdkCode: NEXT_CLOUDINARY_ANALYTICS_ID,
       sdkSemver: NEXT_CLOUDINARY_VERSION,
       techVersion: NEXT_VERSION,