+++ next-cloudinary/src/index.ts
@@ -8,23 +8,32 @@ export { default as CldUploadButton } from './components/CldUploadButton';
 export type { CldUploadButtonProps } from './components/CldUploadButton';
 
 export { default as CldUploadWidget } from './components/CldUploadWidget';
-export type { CldUploadWidgetInfo, CldUploadWidgetProps, CldUploadWidgetPropsChildren, CldUploadWidgetPropsOptions, CldUploadWidgetResults } from './components/CldUploadWidget';
+export type { CldUploadWidgetProps, CldUploadWidgetPropsChildren } from './components/CldUploadWidget';
+export type {
+  CloudinaryUploadWidgetOptions,
+  CloudinaryUploadWidgetResults,
+  CloudinaryUploadWidgetInfo,
+  CloudinaryUploadWidgetInstanceMethods,
+  CloudinaryUploadWidgetInstanceMethodCloseOptions,
+  CloudinaryUploadWidgetInstanceMethodDestroyOptions,
+  CloudinaryUploadWidgetInstanceMethodOpenOptions,
+  CloudinaryUploadWidgetInstanceMethodUpdateOptions,
+  CloudinaryUploadWidgetSources,
+  CloudinaryUploadWidgetError
+} from '@cloudinary-util/types';
 
 export { default as CldVideoPlayer } from './components/CldVideoPlayer';
-export type { CldVideoPlayerProps, CldVideoPlayerPropsColors, CldVideoPlayerPropsLogo } from './components/CldVideoPlayer';
+export type { CldVideoPlayerProps, CldVideoPlayerPropsLogo } from './components/CldVideoPlayer';
+export type { CloudinaryVideoPlayer, CloudinaryVideoPlayerOptions, CloudinaryVideoPlayerOptionsColors, CloudinaryVideoPlayerOptionsLogo } from '@cloudinary-util/types';
 
 export { cloudinaryLoader } from './loaders/cloudinary-loader';
 export type { CloudinaryLoader, CloudinaryLoaderLoaderOptions, CloudinaryLoaderCldOptions } from './loaders/cloudinary-loader';
 
-export type { CloudinaryVideoPlayer, CloudinaryVideoPlayerOptions, CloudinaryVideoPlayerOptionsColors, CloudinaryVideoPlayerOptionsLogo } from './types/player';
-
 export { getCldImageUrl } from './helpers/getCldImageUrl';
-// @deprecated GetCldImageUrl
-export type { GetCldImageUrl, GetCldImageUrlOptions, GetCldImageUrlConfig, GetCldImageUrlAnalytics } from './helpers/getCldImageUrl';
+export type { GetCldImageUrlOptions, GetCldImageUrlConfig, GetCldImageUrlAnalytics } from './helpers/getCldImageUrl';
 
 export { getCldOgImageUrl } from './helpers/getCldOgImageUrl';
-// @deprecated GetCldOgImageUrl
-export type { GetCldOgImageUrl, GetCldOgImageUrlOptions } from './helpers/getCldOgImageUrl';
+export type { GetCldOgImageUrlOptions } from './helpers/getCldOgImageUrl';
 
 export { getCldVideoUrl } from './helpers/getCldVideoUrl';
 export type { GetCldVideoUrlOptions, GetCldVideoUrlConfig, GetCldVideoUrlAnalytics } from './helpers/getCldVideoUrl';