+++ next-cloudinary/tests/helpers/getCldImageUrl.spec.js
@@ -1,11 +1,12 @@
+import { vi, describe, it, beforeEach, afterAll, expect } from 'vitest';
+
 import { getCldImageUrl } from '../../src/helpers/getCldImageUrl';
-import { getCldVideoUrl } from '../../src/helpers/getCldVideoUrl';
 
 describe('Cloudinary', () => {
   const OLD_ENV = process.env;
 
   beforeEach(() => {
-    jest.resetModules()
+    vi.resetModules()
     process.env = { ...OLD_ENV };
   });
 
@@ -26,22 +27,6 @@ describe('Cloudinary', () => {
       });
 
       expect(url).toContain(`https://res.cloudinary.com/${cloudName}/image/upload/c_limit,w_100/f_auto/q_auto/turtle`);
-    });
-  });
-
-  describe('getCldVideoUrl', () => {
-    it('should pass', () => {
-      const cloudName = 'customtestcloud';
-
-      process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME = cloudName;
-
-      const url = getCldVideoUrl({
-        src: 'turtle',
-        width: 100,
-        height: 100
-      });
-
-      expect(url).toContain(`https://res.cloudinary.com/${cloudName}/video/upload/c_limit,w_100/f_auto/q_auto/turtle`);
     });
   });
 })