+++ next-cloudinary/tsup.config.ts
@@ -2,13 +2,11 @@ import { defineConfig, Options } from 'tsup'
 import { plugin as CopyAssetsPlugin } from './plugins/copy-assets';
 
 const commonConfig: Options = {
-  minify: true,
-  target: 'es2018',
-  external: ['react'],
   dts: true,
+  external: ['react'],
   format: ['esm', 'cjs'],
+  minify: true,
   sourcemap: true,
-  clean: true,
 };
 
 export default defineConfig([
@@ -16,5 +14,5 @@ export default defineConfig([
     ...commonConfig,
     entry: ['src/index.ts'],
     esbuildPlugins: [CopyAssetsPlugin] // Add plugin to copy assets on one of the entrypoints
-  },
+  }
 ]);